requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
compress-pickle==2.1.0
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.2.0
filelock==3.13.1
flake8==6.1.0
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.12.2
idna==3.6
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
libaio==0.9.1
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.25.2
packaging==23.1
pathspec==0.12.1
Pillow==10.0.0
platformdirs==4.1.0
pluggy==1.3.0
pyadi-iio==0.0.16
pycodestyle==2.11.1
pyflakes==3.1.0
pylibiio==0.25
pyparsing==3.0.9
pyserial==3.5
pytest==7.4.3
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
zipp==3.16.2
